Implementation bootcamp: 75 stories across catalog, warehouse, access and durable messaging
Commits the previously uncommitted Phase 2 working tree: 25 junior, 30 mid-level
and 20 mid/senior stories, complete with tests, migrations and learning material.

Capabilities added:
- Catalog: tags and collections, draft cloning, variant/gallery editing, import
  staging with preview and commit, category trees, option schemas, sync feed
- Carts and accounts: merge, reusable templates, saved searches, recently viewed,
  checkout preferences, order timelines, reorder, private wishlist notes
- Warehouse: supplier purchase orders and receipts, inventory count sessions,
  bulk atomic stock adjustments, reorder policies, replenishment suggestions,
  packing slips, fulfillment queue, work assignments
- Pricing and returns: read-only checkout quotes, quantity price tiers, shipping
  eligibility, business-day delivery calendars, return eligibility and evidence,
  gift-card ledgers, scheduled discounts
- Access: revocable login sessions, scoped integration API keys, guest order
  access credentials, private account data export
- Messaging: durable webhook outbox with leases and attempt history, historical
  replay, background sales export jobs, cursor-based order history

Scope: 452 files, 60 controllers, 214 endpoints, 46 entities, 26 migrations.

Verified on 2026-09-09 against the frozen source: build succeeds with 0 warnings
and 0 errors; the full suite passes 820/820 with 0 failures and 0 skips in 2m53s.

The implementation predates this commit and was authored in the working tree.
Claude reviewed the tree, verified the build and test results independently, and
performed the commit.

+++ src/Agora.Domain/Entities/ShippingMethod.cs
@@ -42,7 +42,7 @@ public class ShippingMethod
 
     public DateTimeOffset CreatedAt { get; set; } = DateTimeOffset.UtcNow;
 
-    public Money CalculateCharge(Money discountedSubtotal, int totalWeightGrams)
+    public Money CalculateCharge(Money discountedSubtotal, long totalWeightGrams)
     {
         if (totalWeightGrams < 0)
         {
